Jherink

+++ src/Stitch/Elements/SVGSupport.cs
@@ -2,11 +2,11 @@
 {
     public class SVGPoint
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
         public SVGPoint() { }
-        public SVGPoint( int x, int y )
+        public SVGPoint( double x, double y )
         {
             X = x; Y = y;
         }
